Add gradle 2.11 as wrapper
Motivation:
Simplify building of project

Change:
Add gradle 2.11 as wrapper and use it on travis.

Result:
We have control over the gradle version in use and make it easier to execute the build. We can upgrade to gradle 2.13 once we have upgraded the plugins as com.google.protobuf plugin currently in use is incompatible with gradle 2.13.

Fixes: #7